refactor bot logic, output, and add duel mode
This refactors the logic for running the bot into a new bot_logic.rb
file that can be re-used. It also updates how output is formatting,
showing it in the color of the current player and allowing for a prefix.

The big change is the addition of a duel.rb file, that will create a new
game and then spawn two sub-processes each running a bot in the game.
This allows for easily testing the bot logic against itself.